Give the sensor some more time before the overflow happens... 584942 years instead of 72 min.

+++ rosserial_client/src/ros_lib/ros/node_handle.h
@@ -403,7 +403,7 @@ public:
 
   void setNow(Time & new_now)
   {
-    uint32_t mus = hardware_.time_micros();
+    uint64_t mus = hardware_.time_micros();
     sec_offset = new_now.sec - mus / 1000000UL - 1;
     nsec_offset = new_now.nsec - (mus % 1000000UL) * 1000UL + 1000000000UL;
     normalizeSecNSec(sec_offset, nsec_offset);
